Claude/Claude Code 1/requirements.txt
# PDF Intelligence v2 — pinned runtime dependencies
# Install with:  pip install -r requirements.txt
#
# Total uncompressed site-packages target: < 80 MB
# PyInstaller --onefile + UPX target:      < 30 MB

# ── PDF text extraction ──────────────────────────────────────────────────────
pdfplumber==0.11.4          # primary extractor; pulls pdfminer.six + Pillow
pypdf==5.1.0                # pure-Python fallback / encrypted PDFs
Pillow==11.0.0              # already a pdfplumber dep; reused for OCR rendering

# ── Linguistic processing ────────────────────────────────────────────────────
snowballstemmer==2.2.0      # replaces hand-rolled suffix rules (sheet 17)
nltk==3.9.1                 # punkt sentence tokenizer (data downloaded on first run)
                            # regex fallback wired if data absent

# ── Typo-tolerant fallback ───────────────────────────────────────────────────
rapidfuzz==3.10.1           # ~2 MB C-ext; used only when FTS returns 0 hits

# ── EPUB support ─────────────────────────────────────────────────────────────
ebooklib==0.18              # pure-Python EPUB reader
beautifulsoup4==4.12.3      # HTML cleanup inside EPUB chapters
lxml==5.3.0                 # faster BS4 parser (falls back to html.parser if absent)

# ── Optional PDF output from EPUB ────────────────────────────────────────────
reportlab==4.2.5            # pure-Python PDF generation (no wkhtmltopdf needed)

# ── Optional OCR (wrapper only; Tesseract binary installed separately) ───────
pytesseract==0.3.13         # ~50 KB wrapper; Tesseract binary NOT bundled

# ── Dev / packaging (not a runtime dep) ─────────────────────────────────────
# pyinstaller==6.11.1       # uncomment to build .exe
